+++ src/cow-react/modules/limitOrders/updaters/MarketPriceUpdater/hooks/useQuoteRequestParams.ts
@@ -8,7 +8,7 @@ import { getAddress } from '@cow/utils/getAddress'
 import useENSAddress from 'hooks/useENSAddress'
 import { NATIVE_CURRENCY_BUY_ADDRESS } from '@src/constants'
 import { useWalletInfo } from '@cow/modules/wallet'
-import usePrevious from '@src/hooks/usePrevious'
+import usePrevious from 'hooks/usePrevious'
 
 export type LimitOrdersQuoteParams = Omit<FeeQuoteParams, 'validTo'> & { validTo?: number }
