--- src/custom/hooks/useERC20Permit/index.ts
@@ -1,1 +1,0 @@
-export * from './useERC20PermitMod'
