--- src/custom/hooks/useFetchListCallback/index.ts
@@ -1,1 +1,0 @@
-export * from './useFetchListCallbackMod'
