--- src/custom/hooks/useFetchListCallback/useFetchListCallbackMod.ts
@@ -1,42 +1,0 @@
-import { useWalletInfo } from '@cow/modules/wallet'
-import { nanoid } from '@reduxjs/toolkit'
-import { TokenList } from '@uniswap/token-lists'
-import { MAINNET_PROVIDER } from 'constants/networks'
-import getTokenList from 'lib/hooks/useTokenList/fetchTokenList'
-import resolveENSContentHash from 'lib/utils/resolveENSContentHash'
-import { useCallback } from 'react'
-import { useAppDispatch } from 'state/hooks'
-
-import { fetchTokenList } from 'state/lists/actions'
-
-export function useFetchListCallback(): (listUrl: string, sendDispatch?: boolean) => Promise<TokenList> {
-  const dispatch = useAppDispatch()
-
-  // Mod
-  const { chainId } = useWalletInfo()
-
-  // note: prevent dispatch if using for list search or unsupported list
-  return useCallback(
-    async (listUrl: string, sendDispatch = true) => {
-      const requestId = nanoid()
-      // Mod: add chainId
-      sendDispatch && dispatch(fetchTokenList.pending({ requestId, url: listUrl, chainId }))
-      return getTokenList(listUrl, (ensName: string) => resolveENSContentHash(ensName, MAINNET_PROVIDER))
-        .then((tokenList) => {
-          // Mod: add chainId
-          sendDispatch && dispatch(fetchTokenList.fulfilled({ url: listUrl, tokenList, requestId, chainId }))
-          return tokenList
-        })
-        .catch((error) => {
-          console.debug(`Failed to get list at url ${listUrl}`, error)
-          sendDispatch &&
-            // Mod: add chainId
-            dispatch(fetchTokenList.rejected({ url: listUrl, requestId, errorMessage: error.message, chainId }))
-          throw error
-        })
-    },
-    // Mod: add chainId
-    // [dispatch]
-    [chainId, dispatch]
-  )
-}
