+++ src/hooks/useActivityDerivedState.ts
