+++ src/hooks/useAppCode.ts
@@ -1,5 +1,5 @@
 import { useIsGnosisSafeApp } from '@cow/modules/wallet'
-import { DEFAULT_APP_CODE, SAFE_APP_CODE } from '@src/constants'
+import { DEFAULT_APP_CODE, SAFE_APP_CODE } from 'constants/index'
 
 const APP_CODE = process.env.REACT_APP_APP_CODE
