+++ src/hooks/useGetGpPriceStrategy.ts
