+++ src/hooks/useGetSafeInfo.ts
@@ -3,7 +3,7 @@ import { useWeb3React } from '@web3-react/core'
 import { getSafeTransaction } from '@cow/api/gnosisSafe'
 import { SafeMultisigTransactionResponse } from '@gnosis.pm/safe-service-client'
 import { retry, RetryOptions } from 'utils/retry'
-import { RetryResult } from '../../types'
+import { RetryResult } from '../types'
 import { supportedChainId } from 'utils/supportedChainId'
 import { useWalletInfo } from '@cow/modules/wallet'
