+++ src/hooks/useWrapCallback.ts
@@ -7,11 +7,11 @@ import { BigNumber } from '@ethersproject/bignumber'
 import { TransactionResponse } from '@ethersproject/providers'
 import { Contract } from '@ethersproject/contracts'
 import { getChainCurrencySymbols } from 'utils/gnosis_chain/hack'
-import { RADIX_HEX } from '@src/constants'
+import { RADIX_HEX } from 'constants/index'
 import { formatSymbol } from '@cow/utils/format'
 import { getOperationMessage, OperationType } from 'components/TransactionConfirmationModal'
 import { calculateGasMargin } from 'utils/calculateGasMargin'
-import { isRejectRequestProviderError } from '../../utils/misc'
+import { isRejectRequestProviderError } from '../utils/misc'
 import { wrapAnalytics } from 'components/analytics'
 import { useDerivedSwapInfo } from 'state/swap/hooks'
 import { useCloseModals } from 'state/application/hooks'
